requirements.txt
altair==5.0.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
confection==0.1.0
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
cytoolz==0.12.1
decorator==5.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0-py3-none-any.whl#sha256=1618fc62e7aedc2b778caa3ea6e9c8280beac3b83fe7b93316cbb9e8f3abe7c5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
Faker==19.1.0
favicon==0.7.0
floret==0.10.3
fonttools==4.41.0
gitdb==4.0.10
GitPython==3.1.32
htbuilder==0.6.1
idna==3.4
importlib-metadata==6.8.0
jellyfish==1.0.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==9.1.0
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==9.5.0
plotly==5.15.0
preshed==3.0.8
protobuf==4.23.4
pyarrow==12.0.1
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==10.1
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.12.2
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.8.11
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
st-annotated-text==4.0.0
streamlit==1.24.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.5
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
textacy==0.13.0
thinc==8.1.10
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
zipp==3.16.2
fastapi
gunicorn==21.2.0
uvicorn==0.23.2
loguru==0.7.0
python-dotenv==1.0.0
google-generativeai
pymongo==4.4.1
python-multipart==0.0.6
tensorflow==2.13.0
assemblyai==0.16.0
pytube==15.0.0
opencv-python==4.8.0.76
uuid==1.30